refactor(auto-memory): keep bound model and native image helpers
