+++ reme/config/default.yaml
@@ -186,7 +186,7 @@ jobs:
           default: false
         supports_vision:
           type: boolean
-          description: "Caller declares that the selected memory model supports images; does not enable images by itself"
+          description: "Caller declares that the AgentScope wrapper's bound as_llm supports images; does not enable images by itself"
           default: false
         image_mode:
           type: string
@@ -827,10 +827,9 @@ components:
         max_tokens: 65536
         thinking_enable: false
 
-    # Optional vision model for image resources and Auto Memory image inputs.
-    # Direct Auto Memory uses it only for the current image-bearing Agent call;
-    # when absent, it keeps the model already bound to that Agent wrapper.
-    # Image resource/caption processing otherwise falls back to "default".
+    # Optional dedicated vision model for image resources (auto_image_resource_step).
+    # Falls back to the "default" instance above when absent; uncomment to
+    # decouple the vision model from the main LLM.
     # vision:
     #   backend: ${VLM_BACKEND:-openai}
     #   model: ${VLM_MODEL_NAME:-}
